+++ Magic/src/main/java/com/elmakers/mine/bukkit/integration/mobarena/MagicItemThing.java
@@ -1,0 +1,11 @@
+package com.elmakers.mine.bukkit.integration.mobarena;
+
+import org.bukkit.inventory.ItemStack;
+
+import com.garbagemule.MobArena.things.ItemStackThing;
+
+public class MagicItemThing extends ItemStackThing {
+    public MagicItemThing(ItemStack stack) {
+        super(stack);
+    }
+}