+++ Magic/CHANGELOG.md
@@ -14,6 +14,7 @@
  - Mine and Chop can't be cast in no-break regions
  - Add "cost_heading" message option for shops
  - Add "fx" spell for quickly testing particle types
+ - Add more options for disguised entities
 
 # 7.3.6